# api/openapi.yaml
openapi: 3.1.0
info:
  title: Daadno Judicial Services Navigator API
  version: 0.1.0
  description: |
    قرارداد API راهبر خدمات قضایی.

    قواعد سراسری (اجباری، در ENGINEERING_RULES.md تشریح شده):
      1. هیچ اندپوینتی رمز ثنا یا اعتبارنامه سامانه دولتی نمی‌پذیرد.
      2. هر پاسخ تولیدشده باید آرایه `citations` غیرخالی داشته باشد، وگرنه
         باید با `refused: true` برگردد.
      3. هر عدد مهلت فقط از `/cases/{id}/deadlines` می‌آید — هرگز از متن مدل.
      4. کد ملی فقط در اندپوینت‌هایی که صراحتاً آن را لازم دارند پذیرفته می‌شود.
servers:
  - url: https://api.example.ir/v1
...
paths:
  /services:
    get:
      tags: [catalog]
      summary: فهرست خدمات
      parameters:
        - { name: persona,  in: query, schema: { $ref: '#/components/schemas/Persona' } }
        - { name: category, in: query, schema: { type: string } }
        - { name: q,        in: query, schema: { type: string }, description: جست‌وجوی متنی ساده روی عنوان و مترادف‌ها }
        - { name: limit,    in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: cursor,   in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:  { type: array, items: { $ref: '#/components/schemas/ServiceSummary' } }
                  next_cursor: { type: [string, 'null'] }
  /services/{slug}:
    get:
      tags: [catalog]
      summary: کارت کامل خدمت
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceCard' }
        '404': { $ref: '#/components/responses/NotFound' }
  /route:
    post:
      tags: [routing]
      summary: متن آزاد کاربر → خدمات نامزد + پرسش تفکیکی
      description: |
        بدون عوارض جانبی. اگر ابهام باقی باشد `clarifying_questions` پر و
        `candidates` مرتب برمی‌گردد. کلاینت تا رفع ابهام حق ندارد یک نامزد را
        به عنوان پاسخ قطعی نمایش دهد.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:    { type: string, maxLength: 2000 }
                persona: { $ref: '#/components/schemas/Persona' }
                answers:
                  type: object
                  additionalProperties: { type: string }
                  description: پاسخ کاربر به پرسش‌های تفکیکی دور قبل
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [candidates, clarifying_questions]
                properties:
                  candidates:
                    type: array
                    maxItems: 5
                    items:
                      type: object
                      required: [slug, title_fa, score, why]
                      properties:
                        slug:     { type: string }
                        title_fa: { type: string }
                        score:    { type: number }
                        why:      { type: string, description: یک جمله، چرا این نامزد }
                  clarifying_questions:
                    type: array
                    maxItems: 3
                    items:
                      type: object
                      required: [id, question, options]
                      properties:
                        id:       { type: string }
                        question: { type: string }
                        options:  { type: array, items: { type: string } }
                  confident: { type: boolean }
...
  /fees/estimate:
    get:
      tags: [catalog]
      parameters:
        - { name: service_slug, in: query, required: true, schema: { type: string } }
        - { name: khaste_rial,  in: query, schema: { type: integer, description: مبلغ خواسته به ریال } }
      responses:
        '200':
          description: برآورد
          content:
            application/json:
              schema:
                type: object
                required: [items, total_rial, as_of_year, disclaimer]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        label:      { type: string }
                        amount_rial:{ type: integer }
                  total_rial: { type: integer }
                  as_of_year: { type: integer, example: 1405 }
                  disclaimer: { type: string }
  /courts/search:
    get:
      tags: [catalog]
      parameters:
        - { name: q,        in: query, schema: { type: string } }
        - { name: province, in: query, schema: { type: string } }
      responses:
        '200': { description: OK }

  # ---------------- admin ----------------
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    NotFound:
      description: یافت نشد
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: string } }
  schemas:
    Persona:
      type: string
      enum: [haqiqi, hoquqi, vakil, karshenas, danesh_hoquqi]
    Channel:
      type: string
      enum: [self_service, edmat_office, judicial_unit, postal]
...
    ServiceSummary:
      type: object
      required: [slug, title_fa, summary, category, verified_at]
      properties:
        slug:        { type: string }
        title_fa:    { type: string }
        summary:     { type: string }
        category:    { type: string }
        verified_at: { type: [string, 'null'], format: date-time }
        needs_review: { type: boolean, description: true اگر از review_due_at گذشته باشد }
    ServiceCard:
      allOf:
        - $ref: '#/components/schemas/ServiceSummary'
        - type: object
          required: [personas, channels, steps, documents, fees]
          properties:
            aliases:  { type: array, items: { type: string } }
            system:
              type: object
              properties:
                name_fa: { type: string }
                url:     { type: string }
            personas:
              type: array
              items:
                type: object
                properties:
                  persona:          { $ref: '#/components/schemas/Persona' }
                  allowed:          { type: boolean }
                  restriction_note: { type: string }
            channels:
              type: array
              items:
                type: object
                properties:
                  kind:       { $ref: '#/components/schemas/Channel' }
                  is_primary: { type: boolean }
                  deep_link:  { type: string }
                  note:       { type: string }
            prerequisites:
              type: array
              items:
                type: object
                properties:
                  slug:     { type: string }
                  title_fa: { type: string }
                  kind:     { type: string, enum: [hard, soft] }
            steps:
              type: array
              items:
                type: object
                properties:
                  ord:          { type: integer }
                  title:        { type: string }
                  body_md:      { type: string }
                  screen_hint:  { type: string }
                  common_error: { type: string }
            documents:
              type: array
              items:
                type: object
                properties:
                  name:         { type: string }
                  is_mandatory: { type: boolean }
                  format_note:  { type: string }
            fees:
              type: array
              items:
                type: object
                properties:
                  label:      { type: string }
                  basis:      { type: string }
                  as_of_year: { type: integer }
            related_deadline_rules:
              type: array
              items: { type: string }
            legal_refs:
              type: array
              items:
                type: object
                properties:
                  law_title: { type: string }
                  article:   { type: string }
                  url:       { type: string }
...